refactor: use context API for SDK use
