+++ src/App.tsx
@@ -33,6 +33,7 @@ import backgroundImage from 'assets/background.jpg';
 //- Page Imports
 import { ZNS, Staking } from 'pages';
 import PageContainer from 'containers/PageContainer';
+import { ZnsSdkProvider } from 'lib/providers/ZnsSdkProvider';
 
 // Web3 library to query
 function getLibrary(provider: any): Web3Provider {
@@ -93,12 +94,14 @@ function wrappedApp() {
 				<ChainSelectorProvider>
 					<SubgraphProvider>
 						<Web3ReactProvider getLibrary={getLibrary}>
-							{/* Our Hooks  */}
-							<MvpVersionProvider>
-								<EnlistProvider>
-									<App />
-								</EnlistProvider>
-							</MvpVersionProvider>
+							<ZnsSdkProvider>
+								{/* Our Hooks  */}
+								<MvpVersionProvider>
+									<EnlistProvider>
+										<App />
+									</EnlistProvider>
+								</MvpVersionProvider>
+							</ZnsSdkProvider>
 						</Web3ReactProvider>
 					</SubgraphProvider>
 				</ChainSelectorProvider>
